pyproject.toml
[project]
name = "polymarket-us-init"
version = "0.1.0"
description = "Python scripts demonstrating the Polymarket US retail API and SDK"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "polymarket-us>=0.1.2",
    "python-dotenv>=1.2.1",
]
